requirements.txt
Flask==1.1.2
line-bot-sdk==1.19.0
selenium==3.141.0
beautifulsoup4==4.9.3
requests==2.25.1